数据库实体领域模型
