+++ easyclass-domain/src/main/java/edu/vero/easyclass/domain/OnlineClassTest.java
@@ -12,18 +12,19 @@ public class OnlineClassTest {
     private int testId;
     private Date establishTime;
     private Date deadline;
-    private float score;
-
-    public float getScore() {
-        return score;
-    }
-
-    public void setScore(float score) {
-        this.score = score;
-    }
-
+    private TestRecord testRecord;
     private Set<Question> questions;
     private TeacherArragement arragement;
+
+    @Embedded
+    public TestRecord getTestRecord() {
+        return testRecord;
+    }
+
+    public void setTestRecord(TestRecord testRecord) {
+        this.testRecord = testRecord;
+    }
+
 
     @ManyToOne
     @JoinColumn(name="arrageId")
